#1645: refactored validation rules into its own file
