+++ lib/metadataTypes/MetadataType.js
@@ -14,6 +14,7 @@ import deepEqual from 'deep-equal';
 import pLimit from 'p-limit';
 import Mustache from 'mustache';
 import MetadataTypeInfo from '../MetadataTypeInfo.js';
+import validationsRules from '../util/validations.js';
 
 /**
  * @typedef {import('../../types/mcdev.d.js').BuObject} BuObject
@@ -2782,42 +2783,7 @@ class MetadataType {
      */
     static async validation(validationConfig, item, deployDir) {
         const definition = this.definition;
-        const validations = {
-            noGuidKeys: {
-                failedMsg:
-                    'Please update the key to a readable value. Currently still in GUID format.',
-                /**
-                 * @returns {boolean} true=test passed
-                 */
-                passed: function () {
-                    const key = item[definition.keyField];
-                    if (key) {
-                        const regex =
-                            /^[0-9a-z]{8}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{4}-[0-9a-z]{12}$/i;
-                        return !regex.test(String(key).toLowerCase());
-                    } else {
-                        Util.logger.debug('validation-noGuidKeys: key not found');
-                        return true;
-                    }
-                },
-            },
-            noRootFolder: {
-                failedMsg: 'Root folder not allowed. Current folder: ' + item.r__folder_Path,
-                /**
-                 * @returns {boolean} true=test passed
-                 */
-                passed: function () {
-                    /** @type {string} */
-                    const folderPath = item.r__folder_Path;
-                    if (!folderPath) {
-                        // some types do not support folders
-                        return true;
-                    }
-
-                    return folderPath.includes('/');
-                },
-            },
-        };
+        const validations = await validationsRules(this.definition, item, deployDir);
 
         let rule;
